prestashop integration

+++ app/js/stores/cartStore.js
@@ -10,8 +10,16 @@ const CHANGE_EVENT = 'change';
 let _items = [];
 let _cartVisible = true, _cartEnabled = false;
 
+// Clears the cart, called on logout
+function _clearCart() {
+    window.localStorage.removeItem('cart');
+}
+
 // Init products
 function _init() {
+    if(window.location.search.match(/logout/)) {
+        _clearCart();
+    }
 	const cart = window.localStorage.getItem('cart');
 	if (cart) {
 		_items = JSON.parse(cart);
@@ -20,10 +28,16 @@ function _init() {
 
 // Add product to cart
 function _add(update) {
-	_items.push(update);
-    if (typeof window.localStorage !== 'undefined') {
-        window.localStorage.setItem('cart', JSON.stringify(_items));
-    }
+	//Enforce only one combination in cart at a time
+	let existing = _items.some((e) => {
+		return e.combination.id === update.combination.id;
+	});	
+	if(!existing) {
+		_items.push(update);
+        if (typeof window.localStorage !== 'undefined') {
+            window.localStorage.setItem('cart', JSON.stringify(_items));
+        }
+	}
 }
 
 // Set cart visibility
@@ -60,8 +74,9 @@ let CartStore = assign({}, EventEmitter.prototype, {
 	},
 	getCartTotal: function() {
 		return _items.reduce(function (total, item) {
-      let {product} = item;
-			return total + product.price;
+            let {product} = item;
+            let {price} = product.price ? product.price : 0;
+			return total + price;
 		}, 0).toFixed(2);
 	},
 	getCartVisible: function() {
